+++ src/libstd/rt/mod.rs
@@ -55,10 +55,6 @@ Several modules in `core` are clients of `rt`:
 */
 
 #[doc(hidden)];
-#[deny(unused_imports)];
-#[deny(unused_mut)];
-#[deny(unused_variable)];
-#[deny(unused_unsafe)];
 
 use cell::Cell;
 use clone::Clone;
